--- a/benches/base.rs
+++ b/benches/base.rs
@@ -4,7 +4,6 @@ extern crate rand;
 extern crate base_x;
 
 use bencher::Bencher;
-use base_x::alphabet::Binary;
 use base_x::{encode, decode, Alphabet};
 
 
@@ -36,10 +35,6 @@ fn test_encode<A: Alphabet + Copy>(bench: &mut Bencher, alph: A) {
 }
 
 // Actual benchmarks
-
-fn encode_base2_spec(bench: &mut Bencher) {
-    test_encode(bench, Binary);
-}
 
 // Encode UTF-8
 fn encode_base2_utf8(bench: &mut Bencher) {
@@ -106,7 +101,6 @@ fn decode_base58_ascii(bench: &mut Bencher) {
 }
 
 benchmark_group!(benches,
-    encode_base2_spec,
     encode_base2_ascii, encode_base2_utf8, encode_base16_ascii,
     decode_base2_ascii, decode_base2_utf8, decode_base16_ascii,
     encode_base16_utf8, encode_base58_ascii, encode_base58_utf8,