--- a/src/alphabet.rs
+++ b/src/alphabet.rs
@@ -1,37 +1,22 @@
 use std::collections::HashMap;
 
 use encoder::{AsciiEncoder, Utf8Encoder};
+use decoder::{AsciiDecoder, Utf8Decoder};
+
+use DecodeError;
 
 const INVALID_INDEX: u8 = 0xFF;
 
 pub trait Alphabet {
-    type Lookup: CharLookup;
-
     fn encode(&self, input: &[u8]) -> String;
 
-    /// Get a character from Alphabet at index.
-    ///
-    /// This method can panic if usize doesn't fit in base.
-    fn get(&self, usize) -> char;
-
-    /// Get a byte array of the alphabet. This can be useful
-    /// for ASCII-based alphabets.
-    fn as_bytes(&self) -> &[u8];
-
-    /// Returns a lookup type used to find an index of a char
-    /// in the Alphabet.
-    fn lookup_table(&self) -> Self::Lookup;
-
-    /// Get the base (length in characters) of the Alphabet.
-    fn base(&self) -> usize;
+    fn decode(&self, input: &str) -> Result<Vec<u8>, DecodeError>;
 }
 
 #[derive(Clone, Copy)]
 pub struct Binary;
 
 impl Alphabet for Binary {
-    type Lookup = [u8; 256];
-
     #[inline(always)]
     fn encode(&self, input: &[u8]) -> String {
         let cap = input.len() * 8;
@@ -60,83 +45,30 @@ impl Alphabet for Binary {
     }
 
     #[inline(always)]
-    fn get(&self, index: usize) -> char {
-        index as u8 as char
+    fn decode(&self, _input: &str) -> Result<Vec<u8>, DecodeError> {
+        Ok(Vec::new())
     }
-
-    #[inline(always)]
-    fn as_bytes(&self) -> &[u8] {
-        b"01"
-    }
-
-    #[inline(always)]
-    fn lookup_table(&self) -> Self::Lookup {
-        let mut lookup = [INVALID_INDEX; 256];
-
-        for (i, byte) in self.as_bytes().iter().enumerate() {
-            lookup[*byte as usize] = i as u8;
-        }
-
-        lookup
-    }
-
-    #[inline(always)]
-    fn base(&self) -> usize {
-        2
-    }
-}
-
-pub trait CharLookup: Sized {
-    /// Get the index of the `char` in the Alphabet. If `char`
-    /// is not in the Alphabet return `None`.
-    fn get(&self, char) -> Option<usize>;
 }
 
 impl<'a> Alphabet for &'a [u8] {
-    type Lookup = [u8; 256];
-
     #[inline(always)]
     fn encode(&self, input: &[u8]) -> String {
         AsciiEncoder::encode(*self, input)
     }
 
     #[inline(always)]
-    fn get(&self, index: usize) -> char {
-        self[index] as char
-    }
-
-    #[inline(always)]
-    fn as_bytes(&self) -> &[u8] {
-        *self
-    }
-
-    /// Produces the lookup table matching byte index [0 - 255] to a
-    /// corresponding alphabet byte.
-    ///
-    /// The default implementation will produce the lookup table on
-    /// runtime, and recalculate it every time encoding is invoked.
-    /// Ideally a custom implementation of the `Alphabet` would return
-    /// a `&'static` precalculated table here.
-    #[inline(always)]
-    fn lookup_table(&self) -> Self::Lookup {
+    fn decode(&self, input: &str) -> Result<Vec<u8>, DecodeError> {
         let mut lookup = [INVALID_INDEX; 256];
 
-        for (i, byte) in self.as_bytes().iter().enumerate() {
+        for (i, byte) in self.iter().enumerate() {
             lookup[*byte as usize] = i as u8;
         }
 
-        lookup
-    }
-
-    #[inline(always)]
-    fn base(&self) -> usize {
-        self.len()
+        AsciiDecoder::decode(*self, lookup, input)
     }
 }
 
 impl<'a> Alphabet for &'a str {
-    type Lookup = HashMap<char, usize>;
-
     #[inline(always)]
     fn encode(&self, input: &[u8]) -> String {
         let alphabet: Vec<char> = self.chars().collect();
@@ -144,94 +76,51 @@ impl<'a> Alphabet for &'a str {
     }
 
     #[inline(always)]
-    fn get(&self, index: usize) -> char {
-        self.chars().nth(index).expect("Index will be % base, ergo in alphabet range; qed")
-    }
+    fn decode(&self, input: &str) -> Result<Vec<u8>, DecodeError> {
+        let alphabet: Vec<char> = self.chars().collect();
 
-    #[inline(always)]
-    fn as_bytes(&self) -> &[u8] {
-        self.as_ref()
-    }
-
-    /// Produces the hashmap matching any `char` to it's index in alphabet.
-    #[inline(always)]
-    fn lookup_table(&self) -> Self::Lookup {
-        // this is byte-length, which might or might not be enough,
-        // we might suffer a reallocation at some point.
-        let mut map = HashMap::with_capacity(self.len());
+        let mut map = HashMap::with_capacity(alphabet.len());
 
         for (index, ch) in self.chars().enumerate() {
             map.insert(ch, index);
         }
 
-        map
-    }
-
-    #[inline(always)]
-    fn base(&self) -> usize {
-        self.chars().count()
+        Utf8Decoder::decode(&alphabet, map, input)
     }
 }
 
-impl CharLookup for [u8; 256] {
-    #[inline(always)]
-    fn get(&self, byte: char) -> Option<usize> {
-        match self[byte as u8 as usize] {
-            INVALID_INDEX => None,
-            byte => Some(byte as usize)
-        }
-    }
-}
+// #[cfg(test)]
+// mod test {
+//     use super::{Alphabet, CharLookup};
+//     use std::collections::HashMap;
 
-impl<'a> CharLookup for &'a [u8; 256] {
-    #[inline(always)]
-    fn get(&self, byte: char) -> Option<usize> {
-        match self[byte as u8 as usize] {
-            INVALID_INDEX => None,
-            byte => Some(byte as usize)
-        }
-    }
-}
+//     #[test]
+//     fn lookup_str() {
+//         let alphabet = "abcd";
 
-impl CharLookup for HashMap<char, usize> {
-    #[inline(always)]
-    fn get(&self, ch: char) -> Option<usize> {
-        self.get(&ch).map(|index| *index)
-    }
-}
+//         let lookup: HashMap<char, usize> = alphabet.lookup_table();
 
-#[cfg(test)]
-mod test {
-    use super::{Alphabet, CharLookup};
-    use std::collections::HashMap;
+//         assert_eq!(CharLookup::get(&lookup, 'a'), Some(0));
+//         assert_eq!(CharLookup::get(&lookup, 'b'), Some(1));
+//         assert_eq!(CharLookup::get(&lookup, 'c'), Some(2));
+//         assert_eq!(CharLookup::get(&lookup, 'd'), Some(3));
+//         assert_eq!(CharLookup::get(&lookup, 'e'), None);
+//         assert_eq!(CharLookup::get(&lookup, '7'), None);
+//         assert_eq!(CharLookup::get(&lookup, '$'), None);
+//     }
 
-    #[test]
-    fn lookup_str() {
-        let alphabet = "abcd";
+//     #[test]
+//     fn lookup_bytes() {
+//         let alphabet: &[u8] = b"qwer";
 
-        let lookup: HashMap<char, usize> = alphabet.lookup_table();
+//         let lookup: [u8; 256] = alphabet.lookup_table();
 
-        assert_eq!(CharLookup::get(&lookup, 'a'), Some(0));
-        assert_eq!(CharLookup::get(&lookup, 'b'), Some(1));
-        assert_eq!(CharLookup::get(&lookup, 'c'), Some(2));
-        assert_eq!(CharLookup::get(&lookup, 'd'), Some(3));
-        assert_eq!(CharLookup::get(&lookup, 'e'), None);
-        assert_eq!(CharLookup::get(&lookup, '7'), None);
-        assert_eq!(CharLookup::get(&lookup, '$'), None);
-    }
-
-    #[test]
-    fn lookup_bytes() {
-        let alphabet: &[u8] = b"qwer";
-
-        let lookup: [u8; 256] = alphabet.lookup_table();
-
-        assert_eq!(lookup.get('q'), Some(0));
-        assert_eq!(lookup.get('w'), Some(1));
-        assert_eq!(lookup.get('e'), Some(2));
-        assert_eq!(lookup.get('r'), Some(3));
-        assert_eq!(lookup.get('t'), None);
-        assert_eq!(lookup.get('*'), None);
-        assert_eq!(lookup.get('_'), None);
-    }
-}
+//         assert_eq!(lookup.get('q'), Some(0));
+//         assert_eq!(lookup.get('w'), Some(1));
+//         assert_eq!(lookup.get('e'), Some(2));
+//         assert_eq!(lookup.get('r'), Some(3));
+//         assert_eq!(lookup.get('t'), None);
+//         assert_eq!(lookup.get('*'), None);
+//         assert_eq!(lookup.get('_'), None);
+//     }
+// }