--- a/src/alphabet.rs
+++ b/src/alphabet.rs
@@ -1,7 +1,8 @@
 use encoder::{AsciiEncoder, Utf8Encoder};
 use decoder::{AsciiDecoder, Utf8Decoder};
-
 use DecodeError;
+
+use std::ascii::AsciiExt;
 
 const INVALID_INDEX: u8 = 0xFF;
 
@@ -32,14 +33,21 @@ impl<'a> Alphabet for &'a [u8] {
 impl<'a> Alphabet for &'a str {
     #[inline(always)]
     fn encode(&self, input: &[u8]) -> String {
+        if self.is_ascii() {
+            return self.as_bytes().encode(input);
+        }
+
         let alphabet: Vec<char> = self.chars().collect();
         Utf8Encoder::encode(&alphabet, input)
     }
 
     #[inline(always)]
     fn decode(&self, input: &str) -> Result<Vec<u8>, DecodeError> {
-        let alphabet: Vec<char> = self.chars().collect();
+        if self.is_ascii() {
+            return self.as_bytes().decode(input);
+        }
 
+        let alphabet: Vec<char> = self.chars().collect();
         Utf8Decoder::decode(&alphabet, input)
     }
 }