… Cargo.toml
… src/encoder.rs
… src/lib.rs