This is including #8. This changes the encoding implementation from original CPP algorithm, to using a specialized bigint (that I might want to put into another crate altogether), along with some extra optimizations for remainders.

Without BigUint:
```
test encode_base16 ... bench:      64,736 ns/iter (+/- 3,084)
test encode_base2  ... bench:     264,167 ns/iter (+/- 7,460)
test encode_base58 ... bench:      43,857 ns/iter (+/- 1,702)
```

⚡⚡⚡ With BigUint ⚡⚡⚡
```
test encode_base16 ... bench:       9,712 ns/iter (+/- 890)
test encode_base2  ... bench:      18,540 ns/iter (+/- 1,555)
test encode_base58 ... bench:       7,816 ns/iter (+/- 757)
```