@dignifiedquire 2017-02-04
I ran the benches on base-x-native:

before
```
base-x-native (master) ✗ node benchmark/

Seed: aabe04df97f2d26055ced6ba8ca35e75
  2 tests completed.

  rust_encode x 98,869 ops/sec ±1.28% (9 runs sampled)
  js_encode   x 96,920 ops/sec ±1.95% (8 runs sampled)

  2 tests completed.

  js_decode   x 217,634 ops/sec ±0.92% (8 runs sampled)
  rust_decode x 106,053 ops/sec ±11.27% (9 runs sampled)
```

after
```
 base-x-native (master) ✗ npm run bench

> base-x-native@0.0.1 bench /Users/dignifiedquire/opensource/base-x-native
> node ./benchmark/index.js


Seed: d7e7659d9dbe698cd26e4d060d8d79bb
  2 tests completed.

  rust_encode x 357,834 ops/sec ±2.05% (8 runs sampled)
  js_encode   x 101,617 ops/sec ±1.91% (9 runs sampled)

  2 tests completed.

  js_decode   x 136,995 ops/sec ±39.79% (7 runs sampled)
  rust_decode x 201,431 ops/sec ±33.85% (8 runs sampled)
```